# Optimize comparison kernels