--- a/benches/comparison_kernels.rs
+++ b/benches/comparison_kernels.rs
@@ -44,8 +44,8 @@ where
 
 fn add_benchmark(c: &mut Criterion) {
     let size = 65536;
-    let arr_a = create_primitive_array::<f32>(size, DataType::Float32, 0.0);
-    let arr_b = create_primitive_array::<f32>(size, DataType::Float32, 0.0);
+    let arr_a = create_primitive_array_with_seed::<f32>(size, DataType::Float32, 0.0, 42);
+    let arr_b = create_primitive_array_with_seed::<f32>(size, DataType::Float32, 0.0, 43);
 
     c.bench_function("eq Float32", |b| {
         b.iter(|| bench_op(&arr_a, &arr_b, Operator::Eq))