@Dandandan 2021-03-21
There is also a strange part in the benchmark here.

```
    let arr_a = create_primitive_array::<f32>(size, DataType::Float32, 0.0);
    let arr_b = create_primitive_array::<f32>(size, DataType::Float32, 0.0);
```

As the left & right arrays are using the same seed, they have exactly the same elements!